Moved get_transactionsfrom / to() to models.py. Add functions to get a block by seq or hash from blockchain. #42 #49
