+++ exchange_app/api_1_0/blockchain.py
@@ -203,59 +203,29 @@ def get_block_range(start_block, end_block):
     return result.json()['blocks']
      
 
-    
-def get_transactions_from(address, afterhash):
+def get_block_by_hash(hash):
     """
-    return all transactions from address after the one specified by afterhash
-    """
-    
-    #TODO: Read this from blockchain
-    
-    transfers = [
-        {"operationId": "guid", #TODO: Where to get this. If is only valid for this app's transactions, when do we generate/store it? Can blockchain provide it?
-         "timestamp": "20071103T161805Z", #TODO: confirm if should use ISO-8601 basic or extended timestamp representation
-         "fromAddress": address,
-         "toAddress": "xxxxxx",
-         "assetId": "skycoin",
-         "amount": "1000000",
-         "hash": "qwertyasdfg"
-        }, 
-        {"operationId": "guid", 
-         "timestamp": "20180215T231403Z", 
-         "fromAddress": address,
-         "toAddress": "xxxxxx",
-         "assetId": "skycoin",
-         "amount": "2000000",
-         "hash": "asdfgzxcvb"
-        }
-    ]
-    
-    return transfers
-    
-def get_transactions_to(address, afterhash):
-    """
-    return all transactions to address after the one specified by afterhash
+    returns the blocks from blockchain in the specified range
     """
     
-    #TODO: Read this from blockchain
+    values = {"hash": hash}
+    result = requests.get(form_url(base_url, "/block"), params=values)
     
-    transfers = [
-        {"operationId": "guid", #TODO: Where to get this. If is only valid for this app's transactions, when do we generate/store it? Can blockchain provide it?
-         "timestamp": "20071103T161805Z", #TODO: confirm if should use ISO-8601 basic or extended timestamp representation
-         "fromAddress": "xxxxxx",
-         "toAddress": address,
-         "assetId": "skycoin",
-         "amount": "1000000",
-         "hash": "qwertyasdfg"
-        }, 
-        {"operationId": "guid", 
-         "timestamp": "20180215T231403Z", 
-         "fromAddress": "xxxxxx",
-         "toAddress": address,
-         "assetId": "skycoin",
-         "amount": "2000000",
-         "hash": "asdfgzxcvb"
-        }
-    ]
+    if not result.json:
+        return {"status": 500, "error": "Unknown server error"}
+        
+    return result.json()
     
-    return transfers
+    
+def get_block_by_seq(seqnum):
+    """
+    returns the blocks from blockchain in the specified range
+    """
+    
+    values = {"seq": seqnum}
+    result = requests.get(form_url(base_url, "/block"), params=values)
+    
+    if not result.json:
+        return {"status": 500, "error": "Unknown server error"}
+        
+    return result.json()
